Added symbol generation to the script. Using the slimed down tracefiles.

+++ dep-symbol.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+#!/usr/bin/env python3.6
 
 import subprocess
 import os.path
@@ -9,7 +9,10 @@ from optparse import OptionParser
 
 import json
 
+ARCH='x86_64-linux-gnu'
 working_dir = ""
+
+# TODO : Stop packages from re-loading meta
 
 # Really just for tracking a bit more about a symbol stored in our table
 class Symbol:
@@ -72,26 +75,51 @@ def download_deps(deps):
 
     return debs
 
-def extract_debs(debs):
-    # Create some metadata about our little repository
-    meta = []
+def build_symbols(meta):
+    try:
+        subprocess.check_call(['dpkg', '-x', meta.package_deb, 'tmp']) 
+        subprocess.check_call(['dpkg-gensymbols', '-v0', '-p' + meta.package_name, '-etmp/lib/'+ARCH+'/lib*.so*', '-Osymbols'])
+        meta.has_symbols = True
+        subprocess.check_call(['rm', '-rf', 'tmp'])
+    except subprocess.CalledProcessError as err:
+        print(err)
+        print('failed to build symbols file for ' + meta.package_deb)
+    
 
+def extract_debs(debs,metas):
+    # Create some metadata about our little repository
     home = os.getcwd()
 
     for dep,deb in debs.items():
         debhome = os.path.join(working_dir,dep)
+
+        if os.path.exists(debhome):
+            continue
+
         os.mkdir(debhome)
         os.rename(os.path.join(working_dir,deb),os.path.join(debhome,deb))
         os.chdir(debhome)
+
+        print('Extracting ' + deb)
         out = subprocess.check_output(['ar', '-xv', deb])
-        out = subprocess.check_output(['tar', 'xf', 'control.tar.xz'])
+
+        if os.path.exists("control.tar.xz"):
+            out = subprocess.check_output(['tar', 'xf', 'control.tar.xz'])
+        elif os.path.exists("control.tar.gz"):
+            out = subprocess.check_output(['tar', '-xzf', 'control.tar.gz'])
+
 
         # Test for symbol file
         has_sym = os.path.exists('symbols')
+        meta = Meta(dep, deb, has_sym, [])
+        if not has_sym:
+            print(dep + ' has no symbols file. Attempting to build...')
+            build_symbols(meta)
+
+        metas[deb] = meta
 
         os.chdir(home) 
-        meta.append(Meta(dep, deb, has_sym, []))
-        save_meta()
+
 
 
 def parse_symbols(meta,symbols):
@@ -118,18 +146,24 @@ def parse_symbols(meta,symbols):
         
 
 def load_meta():
-    with open(os.path.join(working_dir, 'meta.json'), 'r') as f:
-        meta = json.load(f, object_hook=Meta.as_meta) 
-    return meta        
+    metas = {}
+    if os.path.exists(os.path.join(working_dir, 'meta.txt')):
+        with open(os.path.join(working_dir, 'meta.txt'), 'r') as f:
+            for line in f:
+                m = json.loads(line, object_hook=Meta.as_meta) 
+                metas[m.package_name] = m
 
-def save_meta():
-    with open(os.path.join(working_dir,'meta.json'), 'w') as f:
-        json.dump(meta, f, indent=2, cls=MetaEncoder)
+    return metas
 
-def load_symbols(meta):
+def save_meta(meta):
+    with open(os.path.join(working_dir,'meta.txt'), 'w') as f:
+        for k,m in metas.items():
+            f.write(json.dumps(m, cls=MetaEncoder) + '\n')
+
+def load_symbols(metas):
     symbols = {}
 
-    for m in meta:
+    for k,m in metas.items():
         if m.has_symbols:
             parse_symbols(m, symbols)
 
@@ -141,17 +175,16 @@ def load_trace(name):
         with open(name, 'r') as f:
             for line in f:
                 j = json.loads(line)
-                if j['inst'] == 'call':
-                    calls.append(j)
+                calls.append(j)
     else:
         return {}
 
     return calls
 
-def check_deps(meta,symbols,calls):
+def check_deps(metas,symbols,calls):
     track = {}
-    for m in meta:
-        track[m.package_name] = False
+    for k,m in metas.items():
+        track[m.package_name] = None
 
     for c in calls:
         if c["indirect"]:
@@ -159,22 +192,22 @@ def check_deps(meta,symbols,calls):
         fname = c["fnptr"][1:]
         sym = symbols.get(fname)
         if sym is not None:
-            track[sym.metas[0].package_name] = True
+            track[sym.metas[0].package_name] = sym.libs[0]
         
     # Just for nice output
     used = []
     notused = []
 
     for d,t in track.items():
-        if t:
-            used.append(d)
+        if t is not None:
+            used.append({"package_name":d, "shared_lib":t})
         else:
             notused.append(d)
 
     print('Package has ' + str(len(track)) + ' tracked dependencies')
     print('Using ' + str(len(used)) + ':')
     for d in used:
-        print('\t' + d)
+        print('\t' + d["package_name"] + ' ===> ' + d["shared_lib"])
 
     print('Not using ' + str(len(notused)) + ':')
     for d in notused:
@@ -191,6 +224,8 @@ parser.add_option('-l', '--load', action='store_true', help='jump straight to lo
 
 working_dir = options.working_dir
 
+metas = load_meta()
+
 if len(args) < 1:
     print("error: must supply dependency-list")
     parser.print_usage()
@@ -199,15 +234,14 @@ if len(args) < 1:
 if not options.load:
     deps=read_dependency_list(args[0])
     debs=download_deps(deps)
-    extract_debs(debs)
+    extract_debs(debs,metas)
 
-meta = load_meta()
-symbols = load_symbols(meta)
-save_meta()
+symbols = load_symbols(metas)
+save_meta(metas)
 
 if options.trace is not None:
     calls = load_trace(options.trace)
-    check_deps(meta,symbols,calls) 
+    check_deps(metas,symbols,calls)
